Final switch_gen script and switch statements added to Processor
The final switch statements included 2 errors. Argument passed to
one of the ret calls has to be changed to r.F.getC() and one
instruction needs an erroneous '.read()' removing. The Processor
class in this commit has these corrected. From this point on
changes will be made direct to the java classes (yay!).